add change_word function
Here is a little take away lesson.

Code is all about trying to make it as reusable as possible.
You learn this later when it comes to classes and objects.


So always try and make it as flexible as possible.
So an improvement could be that you can pass the file name to the method as well as the word you want to replace and the word you want to replace it with.
You could make sure that the file is in the same location as the script. Minimizing the need for a full path.

That way you have code that can read any text file and replace any word. Not being limited to alice in wonderland 

Just a thought and you dont have to do it.